chore: split history metadata helpers
